- Changed all physics related stuff to use doubles instead of floats. - Fixed rotation code in PhysicsObject update - Added code to do soft shadow sampling using a kernel during PCF - Also added some code to counteract z-fighting for surfaces that are at an angle relative to the light ray.

+++ physics_engine/physics_engine/AABB.cpp
@@ -20,12 +20,12 @@ void AABB::print()
     std::cout << "AABB; dim = [" << mDim.x << ", " << mDim.y << ", " << mDim.z << "]" << std::endl;
 }
 
-glm::vec3 AABB::getDim()
+glm::dvec3 AABB::getDim()
 {
     return mDim;
 }
 
-void AABB::setDim(glm::vec3& dim)
+void AABB::setDim(glm::dvec3& dim)
 {
     mDim = dim;
 }
