+++ physics_engine/physics_engine/ShadowMap.cpp
@@ -7,12 +7,12 @@
 
 ShadowMap::CameraDirection ShadowMap::sCamDir[6] =
 {
-    { GL_TEXTURE_CUBE_MAP_POSITIVE_X, glm::vec3(1.0f, 0.0f, 0.0f), glm::vec3(0.0f, -1.0f, 0.0f) },
-    { GL_TEXTURE_CUBE_MAP_NEGATIVE_X, glm::vec3(-1.0f, 0.0f, 0.0f), glm::vec3(0.0f, -1.0f, 0.0f) },
-    { GL_TEXTURE_CUBE_MAP_POSITIVE_Y, glm::vec3(0.0f, 1.0f, 0.0f), glm::vec3(0.0f, 0.0f,  1.0f) },
-    { GL_TEXTURE_CUBE_MAP_NEGATIVE_Y, glm::vec3(0.0f, -1.0f, 0.0f), glm::vec3(0.0f, 0.0f, -1.0f) },
-    { GL_TEXTURE_CUBE_MAP_POSITIVE_Z, glm::vec3(0.0f, 0.0f, 1.0f), glm::vec3(0.0f, -1.0f, 0.0f) },
-    { GL_TEXTURE_CUBE_MAP_NEGATIVE_Z, glm::vec3(0.0f, 0.0f, -1.0f), glm::vec3(0.0f, -1.0f, 0.0f) }
+    { GL_TEXTURE_CUBE_MAP_POSITIVE_X, glm::dvec3(1.0f, 0.0f, 0.0f), glm::dvec3(0.0f, -1.0f, 0.0f) },
+    { GL_TEXTURE_CUBE_MAP_NEGATIVE_X, glm::dvec3(-1.0f, 0.0f, 0.0f), glm::dvec3(0.0f, -1.0f, 0.0f) },
+    { GL_TEXTURE_CUBE_MAP_POSITIVE_Y, glm::dvec3(0.0f, 1.0f, 0.0f), glm::dvec3(0.0f, 0.0f,  1.0f) },
+    { GL_TEXTURE_CUBE_MAP_NEGATIVE_Y, glm::dvec3(0.0f, -1.0f, 0.0f), glm::dvec3(0.0f, 0.0f, -1.0f) },
+    { GL_TEXTURE_CUBE_MAP_POSITIVE_Z, glm::dvec3(0.0f, 0.0f, 1.0f), glm::dvec3(0.0f, -1.0f, 0.0f) },
+    { GL_TEXTURE_CUBE_MAP_NEGATIVE_Z, glm::dvec3(0.0f, 0.0f, -1.0f), glm::dvec3(0.0f, -1.0f, 0.0f) }
 };
